requirements.txt
streamlit==1.0.0
Pillow==8.3.2
numpy==1.21.2
tensorflow==2.6.0
keras==2.6.0